Add repository layer
