+++ api/src/repositories/base-repository.ts
@@ -1,0 +1,15 @@
+import { IDBConnection } from '../database/db';
+
+/**
+ * Base class for repositories.
+ *
+ * @export
+ * @class BaseRepository
+ */
+export class BaseRepository {
+  connection: IDBConnection;
+
+  constructor(connection: IDBConnection) {
+    this.connection = connection;
+  }
+}
